+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.6.1][v1.6.1]
+
+### Added
+- Example implementation for easyCredit payment method.
+- Json extension dependency to composer.json.
+
+### Changed
+- Updated readme file.
+
 ## [v1.6.0][v1.6.0]
 
 ### Added
@@ -129,5 +138,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [v1.3.0]: https://github.com/heidelpay/php-payment-api/compare/v1.2.0...v1.3.0
 [v1.4.0]: https://github.com/heidelpay/php-payment-api/compare/v1.3.0...v1.4.0
 [v1.4.1]: https://github.com/heidelpay/php-payment-api/compare/v1.4.0...v1.4.1
-[v1.5.0]: https://github.com/heidelpay/php-payment-api/compare/v1.4.0...v1.5.0
+[v1.5.0]: https://github.com/heidelpay/php-payment-api/compare/v1.4.1...v1.5.0
 [v1.6.0]: https://github.com/heidelpay/php-payment-api/compare/v1.5.0...v1.6.0
+[v1.6.1]: https://github.com/heidelpay/php-payment-api/compare/v1.6.0...v1.6.1